refactor: rename TitleHeader component to HeaderTitle
