+++ slash/css/src/slash.scss
@@ -23,7 +23,7 @@
 @use "./Layout/Header/Logo/Logo";
 @use "./Layout/Header/Name/Name";
 @use "./Layout/Header/NavBar/NavBar";
-@use "./Layout/Header/TitleHeader/TitleHeader";
+@use "./Layout/Header/HeaderTitle/HeaderTitle";
 @use "./Layout/Header/User/User";
 @use "./Steps/Steps";
 @use "./Form/File/File";
