+++ slash/react/src/Layout/Header/HeaderTitle/HeaderTitle.mdx
@@ -1,23 +1,23 @@
 import { Controls, Canvas, Meta } from "@storybook/addon-docs";
-import * as TitleHeaderStories from "./TitleHeader.stories";
+import * as HeaderTitleStories from "./HeaderTitle.stories";
 
-<Meta of={TitleHeaderStories} />
+<Meta of={HeaderTitleStories} />
 
-# TitleHeader
+# HeaderTitle
 
 By default, the title bar is sticky on the top when you scroll down the page. If you want to disable that behavior you can set to false the `isSticky` property of the component.
 
 ## Import
 
 ```tsx
-import { TitleHeader } from "@axa-fr/design-system-slash-react";
+import { HeaderTitle } from "@axa-fr/design-system-slash-react";
 ```
 
 ## Use
 
 ```tsx
 const TitleComponent = () => (
-  <TitleHeader
+  <HeaderTitle
     isSticky
     title="Toolkit Axa"
     subtitle="Info complémentaire"
@@ -31,10 +31,10 @@ export default TitleComponent;
 
 ### Default Title
 
-<Canvas of={TitleHeaderStories.DefaultTitle} />
-<Controls of={TitleHeaderStories.DefaultTitle} />
+<Canvas of={HeaderTitleStories.DefaultTitle} />
+<Controls of={HeaderTitleStories.DefaultTitle} />
 
 ### Complex Title
 
-<Canvas of={TitleHeaderStories.ComplexTitle} />
-<Controls of={TitleHeaderStories.ComplexTitle} />
+<Canvas of={HeaderTitleStories.ComplexTitle} />
+<Controls of={HeaderTitleStories.ComplexTitle} />
